==> MOnSTER/notebooks/1_feature_calculation.ipynb <==
In this notebook a test is performed

import pandas as pd
pd.set_option('display.max_columns', None)
import numpy as np
import sys
import os

from Bio import SeqIO
from Bio.Seq import Seq

from monster.import_data import *

from monster.features_extraction import MotifProperties
from monster.features_extraction import from_lst_to_dict
from monster.features_extraction import feature_calculation

%load_ext autoreload
%autoreload 2

# Data loading

### importing data
#

# positive dataset
seqs_path_pos = '../data/datasets/minc_nr_positive_dataset.fasta'
pos_dict = import_fasta_sequences_as_dict(seqs_path_pos)

# negative dataset
seqs_path_neg = '../data/datasets/minc_nr_negative_dataset.fasta'
neg_dict = import_fasta_sequences_as_dict(seqs_path_neg)

# motifs
lst_motifs = import_list_motifs('../data/lst_motifs')

# Motifs clustering based of physicochemical properties

## Features extraction from the datasets sequences

### Open motifs files for each tool
#
motif_merci = pd.read_csv('../data/merci_lst_motifs.txt',
                          header=None,
                          names=['motif'])

motif_streme = pd.read_csv('../data/lst_motifs_fimo_streme.txt',
                          header=None,
                          names=['motif'])

motif_dimotif = pd.read_csv('../data/dimotif_lst_motifs.txt',
                          header=None,
                          names=['motif'])

### Add a 'tool' column to keep track
#
motif_merci['tool'] = 'merci'
motif_streme['tool'] = 'streme'
motif_dimotif['tool'] = 'dimotif'

print('dimotif', motif_dimotif.shape[0])
print('streme', motif_streme.shape[0])
print('merci', motif_merci.shape[0])


## Features extraction 

### test 

dict_motifs = from_lst_to_dict(lst_motifs)

df_motifs_features = feature_calculation(dict_motifs)
df_motifs_features

len(df_motifs_features)

df_motifs_features.to_csv('df_motifs_features.tsv', header= True,
                         index = None)

df_pos_features = feature_calculation(pos_dict)
df_pos_features

df_pos_features.to_csv('df_pos_features.tsv', header= True,
                         index = None)

df_neg_features = feature_calculation(neg_dict)
df_neg_features

df_neg_features.to_csv('df_neg_features.tsv', header= True,
                         index = None)
